feat: update auth.js v5
