+++ apps/client-fnp/app/api/auth/[...nextauth]/route.ts
@@ -1,2 +1,3 @@
+// export const dynamic = 'force-dynamic' 
 import { handlers } from "@/auth"
 export const { GET, POST } = handlers
