+++ apps/client-fnp/auth.ts
@@ -1,22 +1,24 @@
-import NextAuth, { AuthError, User } from "next-auth";
-import { authConfig } from "@/auth.config";
+import NextAuth, {  User } from "next-auth";
 import Credentials from 'next-auth/providers/credentials'
 import axios, { isAxiosError } from "axios"
 import jwt_decode from "jwt-decode"
+
 declare module "next-auth" {
     interface User {
-        admin: boolean;
-        banned: boolean;
-        exp: number;
-        iat: number;
-        iss: string;
-        subject: string;
-        username: string;
+        admin: boolean
+        banned: boolean
+        exp: number
+        iat: number
+        iss: string
+        subject: string
+        username: string
         token: string
+        email?: string | null
+        emailVerified?: Date | null
     }
 
     interface Session {
-        user?: User
+        user?: User | null
         access_token?: string
     }
 
@@ -26,7 +28,13 @@ import { BaseURL } from "@/lib/schemas"
 
 
 export const { auth, signIn, signOut, handlers } = NextAuth({
-    ...authConfig,
+    debug: true,
+    session: {
+        strategy: "jwt"
+    },
+    pages: {
+        signIn: "/login"
+    },
     providers: [
         Credentials({
             async authorize(credentials) {
@@ -54,8 +62,28 @@ export const { auth, signIn, signOut, handlers } = NextAuth({
                     return null
                 }
             }
-        })],
+        })
+    ],
     callbacks: {
+        async signIn({ user }) {
+            return true
+        },
+        async redirect({ url, baseUrl }) {
+
+            const newURL = new URL(url)
+            const entity = newURL.searchParams.get('entity')
+            const wantToSee = newURL.searchParams.get('wantToSee')
+
+            if (wantToSee !== null && entity !== null ) {
+                return `${newURL.origin}/${entity}/${wantToSee}`
+            }
+
+            // Allows relative callback URLs
+            if (url.startsWith("/")) return `${baseUrl}${url}`
+            // Allows callback URLs on the same origin
+            else if (new URL(url).origin === baseUrl) return url
+            return baseUrl
+        },
         async jwt({ token, user }) {
 
             if (user) {
@@ -69,7 +97,11 @@ export const { auth, signIn, signOut, handlers } = NextAuth({
             if (token.user) {
 
                 const user = token.user as User
-                session.user = user
+                const id = user.id as string
+                const name = user.name as string
+                const email = user.email != null || user.email != undefined ? user.email : ''
+                const emailVerified = user.emailVerified != null || user.emailVerified != undefined ? user.emailVerified : null
+                session.user = { ...user, id, name, email, emailVerified}
                 session.access_token = user.token
             }
