+++ apps/client-fnp/package.json
@@ -36,7 +36,7 @@
     "libphonenumber": "^0.0.10",
     "lucide-react": "^0.279.0",
     "next": "^14.2.3",
-    "next-auth": "5.0.0-beta.4",
+    "next-auth": "^5.0.0-beta.18",
     "next-themes": "^0.2.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
